requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
lxml==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
